% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_jsonl.R
\name{to_jsonl}
\alias{to_jsonl}
\title{Save jsonl file from data.frame}
\usage{
to_jsonl(data, file)
}
\arguments{
\item{data}{Data that can be converted into json file. See ?jsonlite::toJSON()}

\item{file}{A file path to save file to}
}
\value{
File path string of new jsonl file
}
\description{
Save jsonl file from data.frame
}
